+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0]
+
+### Added
+
+- B2500: Add cell voltage sensors
+- B2500: Add overall battery voltage and current sensors
+- B2500: Add calibration information sensors
+- Venus: Add wifi name sensor
+
 ## [1.0.7]
 
 ### Added